Added CORS and JWT support.
The addition of CORS support allows for different frontends that
originate from outside of this app (the Angular and Android frontends)
to make valid cross origin requests.

The JWT support allows for an authentication method that does not
require the frontends to store user auth details (user and pass) locally
and instead just send directly what the user inputs when requesting a
new token. This method also allows for the ignoring of CSRF validation;
the built-in Basic and Session auth in Django require CSRF validation,
JWT doesn't.