Updated rest api
Rest api now displays all the entries by default instead of splitting it over pages. This causes all the info to go to the frontend
Feature added - limit and offset tags can now be used in the url to specify which elements to display. More info about this can be found in FAQ